// src/domain/interfaces/repositories/_gen/ActionResultAppendSheetToSpreadsheetRepository.ts
// This file is generated by scripts/generateInterfaceRepositories.ts
import { ActionResult } from '../../../entities/ActionResult'
import { ActionResultAppendSheetToSpreadsheet } from '../../../entities/ActionResultAppendSheetToSpreadsheet'
import { PromisedResult, UnknownRuntimeError, AlreadyExistsError, NotFoundError } from '../../../../_shared/error'

export interface ActionResultAppendSheetToSpreadsheetRepository {
  issueId: () => PromisedResult<string, UnknownRuntimeError>
  getAll: () => PromisedResult<ActionResultAppendSheetToSpreadsheet[], UnknownRuntimeError>
  getRelevant: (text: string, limit: number) => PromisedResult<ActionResultAppendSheetToSpreadsheet[], UnknownRuntimeError>
  create: (entity: ActionResultAppendSheetToSpreadsheet) => PromisedResult<ActionResultAppendSheetToSpreadsheet, UnknownRuntimeError | AlreadyExistsError>
  update: (entity: ActionResultAppendSheetToSpreadsheet) => PromisedResult<ActionResultAppendSheetToSpreadsheet, UnknownRuntimeError | NotFoundError>
  delete: (id: string) => PromisedResult<void, UnknownRuntimeError | NotFoundError>
  getByActionResultId: (ActionResultId: ActionResult['id']) => PromisedResult<ActionResultAppendSheetToSpreadsheet | null, UnknownRuntimeError>
}