// src/domain/interfaces/repositories/_gen/ActionResultCreateSpreadsheetRepository.ts
// This file is generated by scripts/generateInterfaceRepositories.ts
import { ActionResult } from '../../../entities/ActionResult'
import { ActionResultCreateSpreadsheet } from '../../../entities/ActionResultCreateSpreadsheet'
import { PromisedResult, UnknownRuntimeError, AlreadyExistsError, NotFoundError } from '../../../../_shared/error'

export interface ActionResultCreateSpreadsheetRepository {
  issueId: () => PromisedResult<string, UnknownRuntimeError>
  getAll: () => PromisedResult<ActionResultCreateSpreadsheet[], UnknownRuntimeError>
  getRelevant: (text: string, limit: number) => PromisedResult<ActionResultCreateSpreadsheet[], UnknownRuntimeError>
  create: (entity: ActionResultCreateSpreadsheet) => PromisedResult<ActionResultCreateSpreadsheet, UnknownRuntimeError | AlreadyExistsError>
  update: (entity: ActionResultCreateSpreadsheet) => PromisedResult<ActionResultCreateSpreadsheet, UnknownRuntimeError | NotFoundError>
  delete: (id: string) => PromisedResult<void, UnknownRuntimeError | NotFoundError>
  getByActionResultId: (ActionResultId: ActionResult['id']) => PromisedResult<ActionResultCreateSpreadsheet | null, UnknownRuntimeError>
}