// src/domain/interfaces/repositories/_gen/ActionResultReportResultRepository.ts
// This file is generated by scripts/generateInterfaceRepositories.ts
import { ActionResult } from '../../../entities/ActionResult'
import { ActionResultReportResult } from '../../../entities/ActionResultReportResult'
import { PromisedResult, UnknownRuntimeError, AlreadyExistsError, NotFoundError } from '../../../../_shared/error'

export interface ActionResultReportResultRepository {
  issueId: () => PromisedResult<string, UnknownRuntimeError>
  getAll: () => PromisedResult<ActionResultReportResult[], UnknownRuntimeError>
  getRelevant: (text: string, limit: number) => PromisedResult<ActionResultReportResult[], UnknownRuntimeError>
  create: (entity: ActionResultReportResult) => PromisedResult<ActionResultReportResult, UnknownRuntimeError | AlreadyExistsError>
  update: (entity: ActionResultReportResult) => PromisedResult<ActionResultReportResult, UnknownRuntimeError | NotFoundError>
  delete: (id: string) => PromisedResult<void, UnknownRuntimeError | NotFoundError>
  getByActionResultId: (ActionResultId: ActionResult['id']) => PromisedResult<ActionResultReportResult | null, UnknownRuntimeError>
}